Enlight-backend-prev/dreambooth-api/requirements.txt
pymongo
flask
google-cloud-storage
python-dotenv